+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+# 0.3.4
+* Added `checked_get_set` method to control for `Checkable` entity if `Checked` component is attached.
+
 # 0.3.3
 * Improved `with_local_auto_id_guard` id prefix concatenation logic to correctly work in recursive situations.
 * Renamed `with_local_auto_id_guard` to `with_add_id_pref`. Old function name deprecated until 0.4 release.